+++ test/TestRegionHistogram.cc
@@ -35,7 +35,7 @@ public:
     }
 
     static bool RegionHistogram(const std::string& image_path, const std::vector<float>& endpoints,
-        CARTA::RegionHistogramData& region_histogram, bool is_annotation = false) {
+        CARTA::RegionHistogramData& region_histogram, const std::string& coordinate = "z", bool is_annotation = false) {
         std::shared_ptr<carta::FileLoader> loader(carta::FileLoader::GetLoader(image_path));
         std::shared_ptr<Frame> frame(new Frame(0, loader, "0"));
         carta::RegionHandler region_handler;
@@ -48,7 +48,7 @@ public:
         }
 
         // Set histogram requirements
-        auto histogram_req_message = Message::SetHistogramRequirements(file_id, region_id);
+        auto histogram_req_message = Message::SetHistogramRequirements(file_id, region_id, coordinate);
         std::vector<CARTA::HistogramConfig> histogram_configs = {
             histogram_req_message.histograms().begin(), histogram_req_message.histograms().end()};
         if (!region_handler.SetHistogramConfigs(region_id, file_id, frame, histogram_configs)) {
@@ -89,6 +89,35 @@ TEST_F(RegionHistogramTest, TestFitsAnnotationRegionHistogram) {
     std::string image_path = FileFinder::FitsImagePath("noise_3d.fits");
     std::vector<float> endpoints = {0.0, 0.0, 0.0, 3.0, 3.0, 3.0, 3.0, 0.0};
     CARTA::RegionHistogramData histogram_data;
-    bool ok = RegionHistogram(image_path, endpoints, histogram_data, true);
+    bool ok = RegionHistogram(image_path, endpoints, histogram_data, "z", true);
     ASSERT_FALSE(ok);
+}
+
+TEST_F(RegionHistogramTest, TestStokesRegionHistogram) {
+    std::string image_path = FileFinder::FitsImagePath("noise_4d.fits"); // Stokes I and Q
+    std::vector<float> endpoints = {1.0, 1.0, 1.0, 4.0, 4.0, 4.0, 4.0, 1.0};
+    std::vector<std::string> coordinates{"z", "Iz", "Qz", "Uz", "Plinearz", "PFlinearz", "Pangle"};
+    std::unordered_map<std::string, int> stokes_axis{{"I", 0}, {"Q", 1}};
+    int expected_num_bins = sqrt(4 * 4); // region bounding box is 4x4
+
+    for (auto& coordinate : coordinates) {
+        CARTA::RegionHistogramData histogram_data;
+        bool ok = RegionHistogram(image_path, endpoints, histogram_data, coordinate);
+        std::string stokes_coord = (coordinate == "z" ? "I" : coordinate.substr(0, coordinate.length() - 1));
+
+        if (stokes_axis.find(stokes_coord) != stokes_axis.end()) {
+            ASSERT_TRUE(ok);
+            ASSERT_TRUE(histogram_data.has_histograms());
+            ASSERT_EQ(histogram_data.stokes(), stokes_axis[stokes_coord]);
+            ASSERT_EQ(histogram_data.histograms().num_bins(), expected_num_bins); // default histogram has 1 bin
+            ASSERT_GT(histogram_data.histograms().bin_width(), 0.0);
+            ASSERT_NE(histogram_data.histograms().first_bin_center(), 0.0);
+            ASSERT_FALSE(std::isnan(histogram_data.histograms().mean()));
+            ASSERT_FALSE(std::isnan(histogram_data.histograms().std_dev()));
+        } else {
+            // Fails for U and computed Stokes
+            ASSERT_FALSE(ok);
+            ASSERT_FALSE(histogram_data.has_histograms());
+        }
+    }
 }